Fixed tests for Windows.
CombineArchiveFactory was constructing the URI for an input archive in a way
that did not work on Windows; it was calling the toString() method on the URI,
which resulted in the presence of unescaped backslashes (\) in the URI. We
resolved the problem by relying on Path.toUri() instead.

Also fixed the same mistake [as well as others] in the tests. Notably, the
CreateNewArchiveTest was set up incorrectly because it was testing a file
created on Unix (with line feeds) against a file that the test was constructing
in a Windows environment (with carriage return and line feeds).