converted to Maven based project
As a convention, the source code files of a Maven based project
should go src/main/java. So I have moved them into the right places.
Also, created a pom.xml file and ignored some stuff.

The other important update is to upgrade jena dependency to the newer version
so as to resolve the issues of parsing OMEX file.