enhance the content of an entry in an archive
This commit has made an addition of the master attribute to the
ArtifactInfo class to indicate a file to be master or not.

The commit also updated the unit tests and added a test to testing
the master attribute. The changes include of renaming the type attribute
to the format one so as to be inline with the specification.